Changed programs into txt files instead of strings in the project code Fixed two issues with weight calculations: -switched int to ulong -weight calculation now considers all basic probs instead of only the ones referenced in the calculation for the prob being output

+++ ProbabilisticAssignmentLanguage/SyntaxTree.cs
@@ -105,7 +105,7 @@ namespace ProbabilisticAssignmentLanguage
     }
     public struct ArithMonad : Exp
     {
-        public int Number;
+        public ulong Number;
     }
     public struct BoolMonad : Exp
     {
@@ -114,7 +114,7 @@ namespace ProbabilisticAssignmentLanguage
     public struct ProbMonad : Exp
     {
         public List<ProbElement> Elements;
-        public int Index;
+        public uint Index;
     }
     public struct ProbElement
     {
